feat: move exported decs inspection to manager and improve class
Move the logic of inspecting exported declaration to the same file as the Manager class,
so there's no need to acquire the modules info + exported decs from outside. Also, improve
the computation logic to allow both the result to be returned as an Observable and also an
async/promise result. Finally, add an option to clear the internal subject (cache), so recomputation
can happen should needed.