refactor: change the Set structure into an array and add checking
Change the Set structure into an array since the overload cases are already taken care. Also, it ensures that there is no problem of having the same name across different scopes. Finally, add a checking to make sure that there was in fact a change after applying the canonical form filter to the functions/methods names.